Add terraform variable to manage chart_version
This is required to manage upgrades of the helm_release. If no value is provided, the
helm provider uses the latest version but it does not upgrade the current installed version
to the latest one. So in order to manage it manually, you can either delete and re-create the
resource or use the newly introduced `chart_version` field to update to the desired chart_version.